smoketest: bridge: T3226: Remove redundant code
In 558e1ca9, we found that the interface created by a test case was not
cleaned up completely, which led to the destruction of smoke test.
But in fact, the test case function retained an invalid deletion
configuration code. We added a repair patch and forgot to delete the invalid code.
 Now we delete it