ssh: T3212: remove RestartPreventExitStatus from systemd unit
When configuring SSH to only run inside a given VRF the system can not start
SSHd on bootup as the Kernel will report EPERM (Operation not permitted) when
loading the VRF BPF program. This returns the exit code 255 which is marked
in the systemd unit file to stop restarting the service forever.

Removing this limitation will restart the SSHd on startup and it will live
inside the VRF till the end of days.

(cherry picked from commit cdbac8f10b470a06aff54832da7f006aa3ed194e)