tunnel: T3030: Use the default TTL value of 0 to correspond to the PMTU option
Before 94f886ab ("tunnel: T3173: path MTU discov..."), TTL used the default value
of 255, so c7d08654 ("tunnel: T3030: Modify the comm...") used the default value
of 255. Later, due to the introduction of the PMTU option, the default value was
changed to 0. Continuing to use the default value of 255 will cause the smoke test
to fail, Therefore, the new definition of TTL is migrated