conf_mode: T3047: fix for ospf virtual-link to function when passive-interface default is set.
When an area is directly connected to the backbone area, then the virtual link is used to connect the areas.
But when the passive-interface default is set then the VLINK disconnects and the network becomes unreachable.

In order to fix the issue, 'VLINK' parameter is added to the file "interface-definitions/protocols-ospf.xml.in"
and allows to set this command "set proto ospf passive-interface-exclude VLINK0"