mirror: T3297: Optimize irrelevant error stack hints
When setting the interface, the rule of port image will always be cleared.
Because the TC command will return an error, when the actual error occurs,
the error stack will display the execution information of
the last error command. This is an error.
Now, we force the return value of these two commands
to be overridden to ensure that they always return 0