configsession: T3259: avoid deadlock when data fills stdout pipe
If the subprocess is producing enough data (in this case showConfig on a
large config file), then the construction:

p = subprocess.Popen(.., stdout=subprocess.PIPE, ..)
p.wait()

will deadlock with the subprocess waiting for data to be consumed, while
the Python process waits for its termination. So consume data, then wait
for termination.