route: static: T2450: add missing "dhcp-interface" route option
As thought in the beginning the dhcp-interface route option can not be
superseeded by the interface option. When a route is installed for a DHCP
interface, that interface is usually a broadcast interface which can not be
used for plain interface-based routes.

The old Vyatta logic was migrated to Python where the current received
next-hop address from the DHCP interface is installed as next-hop address.