smoketest: bgp: T2100: fix "simple" testcase for ebgp-requires-policy
Commit 4bf55f97 ("BGP: T2100: Adding RFC8212 option toggle.") added a CLI option
to enable RFC8212 ebgp-requires-policy checks. The extended smoketests assumed
that this will lead to an FRR configuration line of "bgp ebgp-requires-policy" -
which is not the case as this is a default option and FRR hides default
options from the config.

In order to properly verify this functionality we must conduct the negative test
and ensure the option is not present in the CLI at all.