vrf: ospf: T2271: re-arrange xml include building blocks
In order to fully re-use the XML based OSPF CLI definition for per-VRF routing,
the file structure needs to be reorganized and the common OSPF definition is
moved to its dedicated ospf-common-config.xml.i file, which can then be fully
re-included at the VRF level.

As the amount of include files now has reached a certain amount, this also
introduces "per topic" subdirectories on the filesystem to keep a clean
structure.