vrf: ospf: T2271: when removing the entire process cleanup all interfaces
During assembly of the required config changes we also must move the
interfaces_removed assignemnt to an earlier stage so the value is also populated
when the entire process is removed to cleanup all remaining OSPF process assigned
interfaces.

This was yet not the case and when deleting OSPF I still got my "interface eth0"
with the area key configured.