bgp: T1711: remove ASN tagNode and move to "local-as"
Every time when set configuration bgp, you need set AS number. There is very
less benefit in this system so the AS number is moved from a tagNode level down
to a leafNode with the name "local-as", same as on the neighbor or peer-group
level.

This changes the CLI configuration from:

set protocols bgp 100 neighbor 10.10.1.2 remote-as 200

to

set protocols bgp local-as 100
set protocols bgp neighbor 10.10.1.2 remote-as 200