static: route: arp: T3170: prevent empty node.def from beeing generated
Currently if there are XML nodes defined in multiple places it can and will
happen that one overrides the other. In this case, the "protocols static arp"
definition file overwrites the node.def file in the "protocols static" folder
that had a priority - the priority is essential for l2tpv3 to work.