policy: T3490: lower priority to keep up with Policy-Based-Routing (PBR)
PBR is implemented in vyatta-cfg-firewall and the new policy framework uses one
script to rule them all that is triggered from the main policy level.

As the PBR configuration is a subnode with a lower priority we get an inversion,
thus all policy based configuration which applies to FRR only, is set to
priority 200.

Automated tests did not show any regression.