dhcp-server: T3544: fix FileNotFoundError on first invokation
The problem of using the move() operation over render() is that render will
silently create the directory tree in the background and move() does not.

This means that on first boot when /run/dhcp-server does not exist, move will
fail with a FileNotFoundError.

Instead of using move() we render() the configuration two times, one for
validating it via dhcpd -t and the other time to really apply it to the service.

The performance impact should be little as the config should still be cached in
the system RAM.