revert: "util: T2467: add autosudo as an option to command"
Implementing a wrapper which will - based on the command - add a sudo prefix
to the execution string seemed to be a nice idea but unfortunately it did not
only not get momentum but also the codebase somethis added an implicit "sudo"
call a second time.

This resulted in a call: "sudo sudo systemctl"

The entire functionality was removed again and if an op-mode script requires
root priviledges it must be explicitly called with them - no black magic.