ipsec: T3643: move swanctl.conf to /run
This is the completion of commit 50a742b5 ("IPSec: T3643: Fix path for
swanctl.conf file") that moves the generated swanctl file from non-volatile to
a volatile (tmpfs backed) storage like we do for all out configuration files.

Thus it is ensured after a reboot or service deprecation there are no accidential
leftovers from previous configurations stored on the system.