ifconfig: backport ifconfig framework from 1.4 to support new tunnel options
It is easier to backport the entire vyos.ifconfig library from 1.4 instead of
backporting single pieces which are required to add new feature to the tunnel
interface section.

In addition that both libraries are now back in sync it will become much easier
to backport any other new feature introduced in VyOS 1.4!