smoketest: shim: remove superfluous sleep() in getFRRconfig()
The sleep was intended to handle a FRR issue where the config was/is somehow
now available in vtysh even with the commit was done. This rather feels
like a race-condition and is fixed in the subsequent commit.

(cherry picked from commit e7d841d2854d8e0ebb95cb6f0bd83e84fba3a9fa)