container: T2216: no need to query container status
As VyOS CLI is the only truth for dealing with containers we do not need to
query if a container is running, exists or what so ever. We simply always
restart it if something changes and do not rely on the underlaying Linux
status.

If a users does container stuff under the hood - it will be overridden.