op-mode: T3776: drop "frr" level from "restart frr ospfd|bgpd|staticd" commands
The current command to restart any of the FRR processes is:

vyos@vyos:~$ restart frr
Possible completions:
  <Enter>       Execute the current command
  bfdd          Restart Bidirectional Forwarding Detection daemon
  bgpd          Restart Border Gateway Protocol daemon
  ospf6d        Restart OSPFv3 daemon
  ospfd         Restart OSPFv2 daemon
  ripd          Restart Routing Information Protocol daemon
  ripngd        Restart RIPng daemon
  staticd       Restart Static Route daemon
  zebra         Restart IP routing manager daemon

From a real-life example: Two engineers needed 5 minutes to figure it is under
"restart frr" - that is why this commit drops the artificial "frr" level on the
op-mode commands to restart routing protocol daemons.

It's less intuitive to have "restart frr ospfd" or "restart frr bgpd" compared
to "restart ospf" and "restart bgp" - we have the same for "restart ssh" or
"restart snmp" and not "restart openssh sshd".

This commit also drops the d (daemon) suffix of the op-mode comamands so the
commands align with the VyOS CLI, else there would be a miss-understanding from
ospf6d to ospfv3.