isis: T3779: backport entire 1.4 (current) featureset
As IS-IS is a new feature and the CLI configuration changed from 1.3 -> 1.4
(required by T3417) it makes sense to synchronize the CLI configuration for
both versions. This means backporting the CLI from 1.4 -> 1.3 to not confuse
the userbase already with a brand new feature.

As 1.3.0-epa1 is on the way and should not contain any CLI changes afterwards,
this is the perfect time.