vyos.ifconfig: T3619: only set offloading options if supported by NIC
In the past we always told ethtool to change the offloading settings, even if
this was not supported by the underlaying driver.

This commit will only change the offloading options if they differ from the
current state of the NIC and only if it's supported by the NIC. If the NIC does
not support setting the offloading options, a message will be displayed
for the user:

vyos@vyos# set interfaces ethernet eth2 offload gro
vyos@vyos# commit
[ interfaces ethernet eth2 ]
Adapter does not support changing large-receive-offload settings!