login: T3792: bugfix for usernames containing a hyphen
While migrating to get_config_dict() in commit e8a1c291b1 ("login: radius:
T3192: migrate to get_config_dict()") the user-name was not excluded
from mangling (no_tag_node_value_mangle=True).

This resulted in a username "vyos-user" from CLI to be actually created as
"vyos_user" on the system.

This commit also adds respective Smoketests to prevent this in the future.