op-mode: T3619: bugfix "show interfaces X detail"
Commit 27e53fbc ("op-mode: T3619: bugfix "show interfaces" for VLANs") fixed
the op-mode command for the "show interfaces" operation, but if a user was
interested in all the ethernet or bridge interfaces, the command "show
interfaces <type> detail" did not yield any output.

The filtered_interfaces() function was further generalized to only operate on
base components and call itself recusively if required.