vyos.ifconfig: T2738: do not remove OS assigned IP addresses from interface
When using VRRP on any given interface and performing an action against that
interface - be it even only changing the alias - will trigger a removal of the
VRRP IP address.

The issue is caused by:

  # determine IP addresses which are assigned to the interface and build a
  # list of addresses which are no longer in the dict so they can be removed
  cur_addr = self.get_addr()
  for addr in list_diff(cur_addr, new_addr):

When the script calls into the library - we will drop all IP addresses set on
the adapter but not available in the config dict.

We should only remove the IP addresses marked by the CLI to be deleted!

(cherry picked from commit e80d0aebd691f1a707ab534b4d1340fa0b793e01)