vyos.ifconfig: dhcp: T3300: always re-start dhcp client instead of start
Commit dd2eb5e5686655 ("dhcp: T3300: add DHCP default route distance") changed
the logic on how the DHCP process is going to be started. The systemd unit was
always "started" even if it was already running. It should rather be re-started
to track changes in e.g. the DHCP hostname setting.

(cherry picked from commit 8ba8f0e097527e3aaaf8b395bfc07cce47e2c788)