containers: T2216: bugfix config error on image upgrade
As it takes time for the OS to settle while booting up the first time a
container is started after image upgrade it will fail big time. To prevent this
we try to start the container up to 10 times before we generate a hard error.

This makes error-free image upgrade possible again when using containers.