protocols: static: T3680: do not delete DHCP received routes
An ISC DHCP hook script is used to install the received default route into FRR
by simple calls to vtysh. By moving to frr-reload.py the DHCP default route
was deleted as it was not found in the running config.

This commit checks all interfaces if DHCP is enabled and if so - will dynamically
add the route to the generated FRR configuration.