bfd: T4043: add support for passive mode
Mark session as passive: a passive session will not attempt to start the
connection and will wait for control packets from peer before it begins
replying.

This feature is useful when you have a router that acts as the central node of
a star network and you want to avoid sending BFD control packets you don't
need to. The default is active-mode