bgp: T4069: add "parameters shutdown" CLI option
Administrative shutdown of all peers of a bgp instance. Drop all BGP peers,
but preserve their configurations. The peers are notified in accordance with
RFC 8203 by sending a NOTIFICATION message with error code Cease and subcode
Administrative Shutdown prior to terminating connections.

This global shutdown is independent of the neighbor shutdown, meaning that
individually shut down peers will not be affected by lifting it.