bgp: T3967: add support for conditional advertisement
The BGP conditional advertisement feature uses the non-exist-map or the
exist-map and the advertise-map keywords of the neighbor advertise-map command
in order to track routes by the route prefix.

non-exist-map
=============
* If a route prefix is not present in the output of non-exist-map command, then
  advertise the route specified by the advertise-map command.

* If a route prefix is present in the output of non-exist-map command, then do
  not advertise the route specified by the addvertise-map command.

exist-map
=========
* If a route prefix is present in the output of exist-map command, then
  advertise the route specified by the advertise-map command.

* If a route prefix is not present in the output of exist-map command, then do
  not advertise the route specified by the advertise-map command.

This feature is useful when some prefixes are advertised to one of its peers
only if the information from the other peer is not present (due to failure in
peering session or partial reachability etc).

The conditional BGP announcements are sent in addition to the normal
announcements that a BGP router sends to its peer.

CLI nodes can be found under:
* set protocols bgp neighbor <ip> address-family <afi> conditional-advertisement
* set protocols bgp peer-group <p> address-family <afi> conditional-advertisement