containers: T2216: bugfix host networking on image upgrade
The bug was partially fixed with this commit:
https://github.com/vyos/vyos-1x/commit/358f0b481d8620cad4954e3fe418054b9a8c3ecd

The earlier commit introduced a startup retry (up to 10 times) to allow the OS
to settle before the container is started. However, it only applies if
host networking is NOT used. This change applies the same for containers
where host networking is employed.

Since the retry portion of the code (written in the earlier commit) is now
referenced twice, it has been moved to its own function.