vyos.configdict: T4263: leaf_node_changed() must also honor valueLess CLI nodes
If a valueLess node is added or removed from the CLI, a call to
leaf_node_changed() will not detect it.

If node is valueLess, on change old or new (depending on addition or deletion)
will be {} and is treated as None.

Add handler for this special case where old or new is an instance of a
dictionary but empty.

(cherry picked from commit 149f704a172fb14f16d0ba00ef237b972539492f)