scripts: T4269: node.def generator should automatically add default values
Since introducing the XML <defaultValue> node it was common, but redundant,
practice to also add a help string indicating which value would be used as
default if the node is unset.

This makes no sense b/c it's duplicated code/value/characters and prone to
error. The node.def scripts should be extended to automatically render the
appropriate default value into the CLI help string.

For e.g. SSH the current PoC renders:

$ cat templates-cfg/service/ssh/port/node.def
  multi:
  type: txt
  help: Port for SSH service (default: 22)
  val_help: u32:1-65535; Numeric IP port
  ...

Not all subsystems are already migrated to get_config_dict() and make use of
the defaults() call - those subsystems need to be migrated, first before the new
default is added to the CLI help.