smoketest: ssh: verify login of valid and invalid test user
In order to test for proper system authentication and security setup a new
testcase is added which performs an SSH login and command execution with a
predefined user. The result (output of uname -a) must match the output if the
command is run natively.

We also try to login as an invalid user - this is not allowed to work.