dns: T4343: forwarding timeout should be a config
Makes the powerdns `network-timeout` setting configurable via:
`service dns forwarding network-timeout`.
The powerdns default is 1500ms, VyOS now explicitly sets the same
default value or the configured value so that the setting can have a
readily apparent default in the help, rather than the user having to
know it's powerdns.