policy: T4194: simplify prefix-list duplication checks
Commit 5dafe255d ("policy: T4194: Add prefix-list duplication checks") added
first support for FRR prefix-list duplication checks. FRR does not allow to
specify the same profix list rule multiple times.

  vyos(config)# ip prefix-list foo seq 10 permit 192.0.2.0/24
  vyos(config)# ip prefix-list foo seq 20 permit 192.0.2.0/24
  % Configuration failed.
  Error type: validation
  Error description: duplicated prefix list value: 192.0.2.0/24

There is a VyOS verify() function which simply probed for the prefix, action,
le and ge settings - but as Python has excellent support when comparing data,
this can be as simple as a dictionary comparison using "==".